Re-integrate RustPython parser repository

+++ crates/ruff/src/rules/flake8_pytest_style/rules/imports.rs
@@ -36,7 +36,7 @@ pub fn import(import_from: &Stmt, name: &str, asname: Option<&str>) -> Option<Di
 pub fn import_from(
     import_from: &Stmt,
     module: Option<&str>,
-    level: Option<usize>,
+    level: Option<u32>,
 ) -> Option<Diagnostic> {
     // If level is not zero or module is none, return
     if let Some(level) = level {
